pyproject.toml
[project]
name = "djcore52"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "django-allauth>=65.11.0",
    "django-split-settings>=1.3.2",
    "pillow>=11.3.0",
    "python-decouple>=3.8",
]
